Add Heartbeat Plugin
https://github.com/orca-services/cakephp-heartbeat

The heartbeat plugin is a CakePHP plugin providing an application heartbeat status page. It comes with a few sensors, which are configurable, and can easily be extended to e.g. monitor an API. 

We use it in combination with wormly (https://www.wormly.com/), to keep track of our applications.